% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog.R
\name{mylog}
\alias{mylog}
\title{My log function}
\usage{
mylog(x)
}
\arguments{
\item{x}{a numerical vector of length n}
}
\value{
the log of x
}
\description{
A slower implementation of computing the log of a vector
}
\details{
This uses a for loop.
}
\examples{
x <- abs(rnorm(10))
mylog(x)
log(10)
}
